задание спринта 9, третья попытка. Перенесена бизнес-логика в сервис классы
