+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -5,9 +5,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.UserNotFoundException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
 
+import java.time.LocalDate;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -17,13 +19,23 @@ import java.util.stream.Collectors;
 public class FilmService {
 
     private final FilmStorage filmStorage;
+    private Long id = 0L;
+    private static final LocalDate releaseDate = LocalDate.of(1895, 12, 28);
 
     @Autowired
     public FilmService(@Qualifier("inMemoryFilmStorage") FilmStorage filmStorage) {
         this.filmStorage = filmStorage;
     }
 
+    private Long generateId() {
+        return ++id;
+    }
+
     public Film addFilm(Film film) {
+        if (film.getReleaseDate().isBefore(releaseDate))
+            throw new ValidationException("Attempt to add film " +
+                    "with releaseDate before 28-12-1895");
+        film.setId(generateId());
         return filmStorage.addFilm(film);
     }
